feat: usecases factories
