--- backend/src/main/factories/usecases/auth-login/auth-login.factory.ts
@@ -1,5 +1,0 @@
-import { AuthLoginUsecase } from '@/data/usecases/auth-login/auth-login.usecase'
-
-export const makeAuthLoginUsecase = (): AuthLoginUsecase => {
-  return new AuthLoginUsecase()
-}
