--- backend/src/main/factories/usecases/auth-register/auth-register.factory.ts
@@ -1,5 +1,0 @@
-import { AuthRegisterUsecase } from '@/data/usecases/auth-register/auth-register.usecase'
-
-export const makeAuthRegisterUsecase = (): AuthRegisterUsecase => {
-  return new AuthRegisterUsecase()
-}
