+++ backend/src/main/factories/usecases/auth/auth-refresh/auth-refresh.factory.ts
@@ -1,4 +1,4 @@
-import { AuthRefreshUsecase } from '@/data/usecases/auth-refresh/auth-refresh.usecase'
+import { AuthRefreshUsecase } from '@/data/usecases/auth/auth-refresh/auth-refresh.usecase'
 
 export const makeAuthRefreshUsecase = (): AuthRefreshUsecase => {
   return new AuthRefreshUsecase()
