+++ backend/src/main/factories/usecases/auth/index.ts
@@ -1,0 +1,3 @@
+export * from './auth-login/auth-login.factory'
+export * from './auth-refresh/auth-refresh.factory'
+export * from './auth-register/auth-register.factory'
