--- backend/src/main/factories/usecases/create-feature-flags/create-feature-flags.factory.ts
@@ -1,5 +1,0 @@
-import { CreateFeatureFlagUsecase } from '@/data/usecases/create-feature-flags/create-feature-flags.usecase'
-
-export const makeCreateFeatureFlagUsecase = (): CreateFeatureFlagUsecase => {
-  return new CreateFeatureFlagUsecase()
-}
