--- backend/src/main/factories/usecases/delete-feature-flags/delete-feature-flags.factory.ts
@@ -1,5 +1,0 @@
-import { DeleteFeatureFlagUsecase } from '@/data/usecases/delete-feature-flags/delete-feature-flags.usecase'
-
-export const makeDeleteFeatureFlagUsecase = (): DeleteFeatureFlagUsecase => {
-  return new DeleteFeatureFlagUsecase()
-}
