+++ backend/src/main/factories/usecases/feature-flags/index.ts
@@ -1,0 +1,5 @@
+export * from './create-feature-flags/create-feature-flags.factory'
+export * from './delete-feature-flags/delete-feature-flags.factory'
+export * from './update-feature-flags/update-feature-flags.factory'
+export * from './find-feature-flags/find-feature-flags.factory'
+export * from './/list-feature-flags/list-feature-flags.factory'
