--- backend/src/main/factories/usecases/find-feature-flags/find-feature-flags.factory.ts
@@ -1,5 +1,0 @@
-import { FindFeatureFlagUsecase } from '@/data/usecases/find-feature-flags/find-feature-flags.usecase'
-
-export const makeFindFeatureFlagUsecase = (): FindFeatureFlagUsecase => {
-  return new FindFeatureFlagUsecase()
-}
