--- backend/src/main/factories/usecases/list-feature-flags/list-feature-flags.factory.ts
@@ -1,5 +1,0 @@
-import { ListFeatureFlagsUsecase } from '@/data/usecases/list-feature-flags/list-feature-flags.usecase'
-
-export const makeListFeatureFlagsUsecase = (): ListFeatureFlagsUsecase => {
-  return new ListFeatureFlagsUsecase()
-}
