--- backend/src/main/factories/usecases/update-feature-flags/update-feature-flags.factory.ts
@@ -1,5 +1,0 @@
-import { UpdateFeatureFlagUsecase } from '@/data/usecases/update-feature-flags/update-feature-flags.usecase'
-
-export const makeUpdateFeatureFlagUsecase = (): UpdateFeatureFlagUsecase => {
-  return new UpdateFeatureFlagUsecase()
-}
